pyproject.toml
[tool.poetry]
name = "django-rest-template"
version = "0.1.0"
description = ""
authors = ["Anwar Hossain <mahedianwar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
psycopg = "^3.1.13"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
